handle fund allocation for security

+++ api/routers/contract/adjustSecurityAllocationPercentageRequest.py
@@ -4,4 +4,4 @@ class AdjustSecurityAllocationPercentageRequest(BaseModel):
 
     accountId : str
     securityId : str
-    allocationPercentage : float
+    allocationPercentage : int
